multi_rag_system/pyproject.toml
[project]
name = "multi-rag-system"
version = "0.1.0"
description = "Multi-source RAG system with Gemini and Tavily"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.3.0",
    "google-generativeai>=0.3.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.1.0",
    "tavily-python>=0.1.0",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["main"]
